Changed members of Costs to double, cumulative cost is now the node cost plus the costs of the children divided by their out degree

+++ stratosphere-compiler/src/main/java/eu/stratosphere/compiler/plandump/PlanJSONDumpGenerator.java
@@ -571,11 +571,11 @@ public class PlanJSONDumpGenerator {
 		writer.print("\" }");
 	}
 
-	public static final String formatNumber(long number) {
+	public static final String formatNumber(double number) {
 		return formatNumber(number, "");
 	}
 
-	public static final String formatNumber(long number, String suffix) {
+	public static final String formatNumber(double number, String suffix) {
 		final int fractionalDigits = 2;
 
 		StringBuilder bld = new StringBuilder();
